More robust GeoJSON provider by using geojson library
